Expose flake packages to flake modules
This feels like a hack, but short of making the user place us in
specialArgs as follows we can't access our own packages.
nixpkgs.lib.nixosSystem {
  system = "aarch64-linux";
  specialArgs = {
    inherit nixos-sbc;
  }
  modules = [
    nixos-sbc.nixosModules.default
  ];
}

This is not great as the name a user uses in specialArgs matters to our
ability to reference ourself, as well as putting us in more places that
make it easier for mistakes to occour.

Using callPackage on our pkgs directory seems to do the trick, and so to
not have to do that ad-nausium, we do it in our base sbc module and add
it to our args.