Make the user think about RegDom
This flag is here to make the user have to consider their regulatory
domain before they enable WiFi radios.  This is only for hardware that
is sketchy, such as hardware where the device-specific wifi parameters
are provided via a generic binary blob for all devices.

This is more than what OpenWRT does, which doesn't present any warning
and turns the radios on when you plug it in the very first time.