Add method to silence errors on bootstrap images
The BPiR3 is going to have its wifi set to enabled by default, but
enabling wifi without accepting regdom on this device will result in an
error being thrown.  It's not generic if I need to worry about every
SBCs boot image options, so if we're making a boot image, this will be
used to change defaults for only the bootstrap image.