Add rough board-spec definition to project
This is very early, and certainly going to change.  At the very least
I'm going to need the ability to include the definitions of things such
as uart/i2c/rtc in other places, so they'll be broken out.

Disabling an enabled doesn't do anything right now, and the same goes
for enabling the disabled.  This is just a start.  Disabling/enabling
needs to be somewhat flexable, as you can't always rely on a single
method to solve all problems.  For example, lets pretend RTC DS3231
existed on the board, we could easily disable it with at dtoverlay,
meanwhile if we had a RK808 RTC, we can't disable it, as it's part of
the PMIC, so we'd want to blacklist the kernel module instead.

The spec is probably best placed in the board folder under a new
spec.nix file, but I'll let that be future churn.

The goal will be to progress the spec forward enough that I can
configure the "console" kernel argument from it in a generic way, as
well as get an RTC up.  Once those are done, there should be a
repeatable pattern ready for anything else that comes along, and I can
focus on boot, sd-images, and binary caches.