Build kernel console param from UARTs
While it would be very easy to simply type this in every board, the
names vary, and sometimes the hardware needs to be enabled.  This is
intended to be similar operation for other hardware, such as RTCs.
There's a lot of duplication right now, and it would be good to reduce
that.