Add RTC module
This is the next thing to work out, and in writing this it's clear the
current 'sbc.spec' structure is a failure.  It does not make since to
have sbc.spec.rtc and sbc.rtc, because spec is only for base hardware,
but sbc.rtc would have addon hardware, that is simply confusing.

sbc.console makes since as you might have UART that you don't want a
console on, but it doesn't make since to sit next to sbc.uart at the top
level.  It does make since for uart to have a "console" option however.
That seems likely to work well even with problematic hardware like the
RPi.