Move devices into sbc.board and remove duplicate settings
Adding hardware, such as a ds3231 RTC will now occour in what was the
"spec" before.  It's not a spec anymore, so "board" seems more
appropriate.  As such 'sbc.console' and 'sbc.rtc' don't make much sense,
especially the later, as it would result in duplicate device definitions
and contain spec devices (some, none, or all) as well as new ones,
complicating management.

While non-DT UART could be added and a console desired, such as a USB
device, that's not really a SBC specific thing, so it's rather out of
scope, so keeping console in with UART does seem better too.