Default sbc.board.<device>.devices to an empty set instea of null
If a device is not defined, this will throw an error due to how we
iterate over the devices to see if we need to do anything to them.

By having an empty set instead of it being null the error goes away,
and I don't see any reason explicitly handling the null is better than
having an empty attrset.