Tease single-partition BTRFS layout in README
This is one of the planned defaults, so tease it.  I'm pretty sure uboot
supports comrpessed btrfs partitions, so this would improve SBC
management as /boot won't run out of space so quickly, and compression
tends to be good with SD cards as SD cards are slow, and even with
limited CPU time the CPU time is cheap.

BTRFS is better at detecting corruption as well, which is beneficial
with cheap and/or old SD cards.