Return initrd compressor to gzip
I thought initrd was decompressed by kernel, not u-boot, but I think I'm
wrong about that.  Using the booti command with a zstd initrd resulted
in it complaining about the format.  ZSTD produces an 11MB initrd and GZ
a 13MB initrd, GZ is safe, and 2MB isn't a big deal when the kernel
size is 60-70MB.