U-boot search prefix first
Specify /boot before / in u-boot exlinux boot config so u-boot searches
the most specific path before trying more generic paths.

The initial config of u-boot comes from my private repo, which I quickly
broke out for others to use in nixos-bpir3-example.  Others ran into an
issue I also ran into, that boot fills up if you are not careful, and
the only way out of that is manual intervention that requires
non-trivial nixos knowlage.  This repo aims to do away with the boot
partition and keep everything on the main partition, so changing the
order is in order.