Add tooling to build ext4 bootable BPiR3
This is largely the same tooling as was included in the
nixos-bpir3-example repo, but sd-image-mt7986 has been modified to
work without a /boot partition.  For u-boot to find the partition during
bootflow it needs to either be set to the ESP GUID or have the "Legacy
Bootable" attribute set.  Legacy bootable attribute flag is bit 2, so we
set it on the root partition.

This needs to be made more generic, because copying it for every SBC
model is obviously a non-starter, as is duplicating it for btrfs.