Begin modularising boot image support
Start breaking it out, and adding btrfs support.
BTRFS seems unlikely to work with bootflow currently, so it'll probably
be put on the backburner until I flush out enough that I add a second
SBC, which is going to be something not in production, because taking
the whole network down is not fun.