Update uboot to 2024.01
I don't chase bugs in old versions, it's usually a waste of time.
In this case, the issue is still present.  Bootflow fails to read
/boot/extlinux/extlinux.conf when using bootflow, despite being able to
access it with the ls and load commands on the u-boot command line.

dtc needs to be a path, or build fails with no build target for dtb's.
If CONFIG_DEVICE_TREE_INCLUDES is defined, it fails the same.  There
isn't a -u-boot.dtsi for this board, so for now it's getting dropped
there.