Add pine64 rock64 board without uBoot configured
The rock64 is a really old board with the rk3328 processor.  There were
a number of boards produced with rk3328 processors, and as such upstream
support of the rk3328 is pretty good and we don't have to do much of
anything to have a working board.

The exception to this is the u-boot cmdline, which doesn't work for us
out of the box.  Pretty much every OS distribution does their u-boot
differently, so this is something always configured.

Unlike the BPiR3, there already exists a <board>-u-boot.dtsi file, so
dropping the file into that location isn't going to work.  This will be
tackled next.