Make Rock64 bootable
If we copy our dtsi to include into the tree in the DTC folder and then
CONFIG_DEVICE_TREE_INCLUDES it at the relative path, it gets included
without issue.

CONFIG_BOOTCOMMAND has to be set manually as it is set during _defconfig
and the default is not updated when our extraConfig is pasted over it.