Add device.disableMethod to rtc and i2c
UART and wifi modules are still being ignored for enable/disable options
as they're not yet useful and the duplication takes a lot of effort to
not stuff up.  Really need to make this generic.

The disable method is so the RTC on the Rock64 can be disabled.
As there is no battery access on the v2, and the v3 battery is external,
the RTC for this device will be disabled by default on this device.

Disable is added to i2c as well to keep the two modules from diverging.