Remove btrfs defragment as nix store is already RO
This either has to occur earlier in boot, likely during stage 1, or it
will have to happen with the help of a namespace to allow remounting the
nix store read-write without globally exposing it as such.

It takes a few mintues, so probably late in userspace would be the best
time.

The store will naturally compress over time through use as it is
mounted with the compress flag, so it's not that nesessary to force it.

Also, this didn't actually compress because I forgot the -czstd flag.