Add support for bootstrap with BTRFS subvolume layout
This is the format that is destened to be the default.  Subvolumes are
useful, but they do have a drawback, you can't reflink between /nix and
/boot without mounting the top-most subvolume and operating on it.

Overall, having subvolumes provides a lot of flexibility for what is a
small drawback.

One such example of the flexibility is that you can quickly convert to
using impermanence by including it, switching `/` to a tmpfs, mounting
the top-most btrfs subvol, creating /@persistance, including it in your
fileSystem layout, and rebooting, at which point you can nuke /@.

Normally, I'd call /@, /@root, but some people want to put the root user
on a subvol at /@root, so I'll use the /@ style for the `/` subvol here.