Bump version to 0.2 and set btrfs-subvol as default layout
0.2 is a big jump, as it changes the default disk format.  This is the
desired state, so it's likely future bumps are going to be far less
impactful and bumps that can be done without re-provisioning your
system.

Likely, exposing the disk-format into the raw image will make since, and
therefor btrfs-subvol will be the default, but the version will evolve
to mean something different, more akin to NixOS versions.