Create firmware packaging for RPi4b
If you have an idea of how most SBCs boot, you can forget all of that
because this is a RaspberryPi and it's insane.  Instead of booting a ARM
system with a GPU attached you're booting a GPU with an ARM attached.

The whole first stage of the boot is completely closed source blobs.
But ultimately the FAT "firmware" partition, which used to be the boot
partition, is where this first stage looks for its blobs.  Now unlike
every other system "u-boot.bin" is dtb-less, even though it's not the
"u-boot-no-dtbs.bin".  So how is it booting?  Why it's using the DTB
that firmware read off of the firmware partition modified and then left
in memory for it to use.  This is confusing because upstream uses a
pre-built on, but this means what you expect uboot to do might not be
true.  You can ether use the u-boot one, or build the extras as an
overlay and use the RPi overlay mechanism to add it.  I'll take the
u-boot dtb.

So, how does the CM4 and all that work with this scheme?  Well for
u-boot it's fine, but for NixOS it doesn't, because NixOS always uses
the deterministic dtb for the system, which is good.  But that means
getting it working on a CM4 or other systems will require actual work
to produce the overlays.  What does the firmware change?  I dunno, I
don't think this is documented.  But I guess it's probably picking up
`bcm2711-rpi-cm4.dtb` out of the overlay directory, and the source for
that is available in the vendor kernel.

You probably need to try to boot one of these with earlyuart enabled
to get an idea, as I expect it will fail to find that as I don't have
that here.