Add RaspberyPI 4 board configuration
This produces a working system, but I haven't tested anything like
graphical acceleration yet.  RPi's are a pain in the ass, they'll boot
in one config, and change a kernel and it'll lockup, hopefully that's
not going to happen here.  If it does, drop the console= and loglevel
commands and try again, because that seems to be an area thats really
brittle.