Correct typo of linuxPacakges to linuxPackages
Packages was misspelled, correct that.

How does a misspelling like that happen?  Copy and paste after typing it
once and QWERTY!  Packages is heavily biased to left hand, and contains
a redirect that is easy to get carried away with as it's bouncing back
to the last character hit.
Left hand: A->C->A and you need to hit K between the C and A.

Co-authored-by: Stefan Junker <mail@stefanjunker.de>