Add reformat rev to git blame ignore file
When blame is run after a mass reformat everything will show as being
touched by that commit.  This makes finding the commit that changed
something difficult/slow and make it more difficult to understand why
something was added or changed.

This file can be configured in local git and is used by github to avoid
that pitfall making reformats more tenable.