bpir3: Build ATF and u-boot in usual order
We're a bit of an outlier, building ATF after u-boot, as everything else
builds u-boot after ATF, incorporating ATF blobs into u-boot's output.

This was difficult with MT7986 as all prior art does it the opposite.
The MT7986 does not exist in upstream ATF, so we have to use a vendor
fork, and it's been patched in a way where it wants tools from uboot to
build bl2.bin, so it was easier to just do what it wanted.

This is annoying as working on u-boot is far more likely than hacking on
ATF, but I was too lazy to deal with it then.  With cross-compiling,
dealing with the vendor ATF wanting to call its own in-tree copies of
its tools is a problem.  Rather than hack around that, get our order
right and call fiptool after uboot builds.

Currently this is untested, but fip.bin produced by ATF and u-boot hash
the same, so it should be right.