Improve filesystem resize reliability
If the device being resized is not the first block device, its
partitions may not start with minor 0 for the block device and minor 1
for the first partition.  This results in a bogus partition being
created immedately following the root partition.

`lsblk` has PARTN column, meaning we don't even need to do fragile
parsing of the block device to get this right.  (sda3, vda3, hda3,
nvme0n1p3, mmcblk0p1, etc).