Fix RAMIFY when extreme number of directories exist
When an extreme number of directories exist in a directory tree being
moved between btrfs subvolumes the metadata usage explodes and we fill
up the extremely limited space we operate in during the first boot,
before resize.

With the release of 24.05, an extreme number of directories exist in the
minimal installation by default as nixpkgs is included in the system
closure.

By creating a snapshot of the top-level subvolume we avoid making a new
directory tree.  Then we remove the files from the top-level and empty
out the subvolume of anything extra, before moving the contents of the
mountpoint up into the new root.

Other methods of attempting to deal with this included copying then
removing each store path one at a time, which failed.  `mv` doesn't
work across filesystem boundaries, which this is.

It was a bit of a hisenbug, because it depends on how large metadata is
during mkfs, and some of my images are 512MB, and thus have more than
enough space, and some are 256MB, which would fail.