sbc: Improve modularity of dtOverlay methods
The OrangePi5B will require a bit more than setting status = "okay" to get
overlays working, so the simple-enable method isn't going to cut it
anymore.

simple-enable was very strict, and had it's arguments in a hard-coded
order that is difficult to extend without proliferation of tons and tons
of xyz-enable variants.

By working with submoduleWith, an appropriate amount of info can be
passed up into the submodule, and reasonable defaults can be set, letting
the dtOverlay support be fairly flexible.

This is still messy, and there's a lot of duplication between i2c and
rtc.  Additionally UART needs to grow this same functionality.  These
need to be made modular so there's less code duplication so the code can
be more reliably built upon.