SBC: Reduce duplicated code in dtoverlay and module options
These aren't the hard things to maintain, but it's the low-hanging fruit
and that's a good spot to start I guess.

It might not even stay this way, I'm not sure.  Device is going to get
deduplicated soon, but UART is going to be the device that is going to
challenge all the current work, as it has things that are specific to it
that should actually build on top of everything i2c and RTC has, because
all the enable/disable functionality is missing from UART, but it has a
bunch of its own complexity.