Fix device timeout during RAMIFY with tmpfs root
If an installation has been converted to use a tmpfs root, we spend a
ton of seconds waiting for "none" to appear for moutning.  Once
waitDevice times out, we continue and do nothing.

If `rootDevice` is "none", then skip adding the RAMIFY script to initrd.

Devices should support this layout on initial-boot, but currently that
is not the case, so this does not remove functionality at this time.
This should however be corrected.