Patch u-boot with fix for an unbootable btrfs situation
See patch:
https://lore.kernel.org/u-boot/20240618214138.3212175-1-alexthreed@gmail.com/

The issue presents as such during boot:
> BUG at fs/btrfs/extent-io.c:629/free_extent_buffer_internal()!
> BUG!
> resetting ...

Using the reproduction method outlined here, I used a copy of my
filesystem to check this patch against it, and it results in a bootable
system.
https://lore.kernel.org/u-boot/3281192.oiGErgHkdL@youmu/

I had one system do this 5 months ago, and another one do it yesterday.
The first one I wasn't sure if it wasn't a one-off bug.  The timing of
this patch is very fortunate as otherwise I'd be making ext4 the
default.