BPiR3: Minimal-ish kernel for CI builds
The current kernel used by the BPiR3 is basically a "allmodconfig"
kernel, which means we're building way way too much stuff.  Additionally
there's a lot of other configuration we can get away with turning off.

This takes the build-time down from ~8hr on a OrangePi 5B to ~1hr.  On
the actual BananaPi R3, the build time for the former would probably
stretch into a full day.

My end goal here is going to be to do away with the non-minimal kernel
and make this one the default, however both need to exist for a period
to allow a transition.  It should compile in a reasonable amount even
on the R3, but the goal is to do this with CI and push it into a cache.

If a feature is missing or has been explicitly disabled that you want,
feel free to request it is added, the goal is a reasonable kernel build
not an absolute minimum.