Add _lib.builders.buildTargets as filtered list of drvs
This makes a list of filtered derivations to build in CI.  I want it to
fail if build fails, which will signal there's a problem, so I can't let
things fail for being unfree and just keep going as it would not be very
noticeable.  I also have limited compute that's hanging off a slow
internet pipe, so it needs to kick big things, autoModules kernels, out.

This is a monstrosity and was painful and time consuming to write, I
would very much appreciate ways to make this less horrible.