Add board info files
First pass with primary focus on the BPiR3.

Goal is to document status, such as kernels and patches and what works
and doesn't.

In the long run I want to standarise some testing of various features,
such as what is actually on the expansion bus, do those things work with
upstream, and would they ever work.  I have a board, which is not in
this repository, that looks good on paper until you realise that to
enable X feature on the expansion bus you need to turn off the I2C
interface that's connected to the PMIC, so you effectivly cannot ever
use those pins on the expansion bus or X as it is only available on
those pins.