u-boot: Add output hash to version
While we are not auto-upgrading u-boot, simply knowing the u-boot
version isn't that informative.  By including the hash used in $out we
can descriptively identify the u-boot in use.

I don't know if it's possible to easily go from hash to source tree,
but it certainly can be used to identify if it needs to be updated and
also tell you if the u-boot used to boot is the one you expect it to be,
which may not always be the case if there's a second u-boot in a
higher-priority location.

Example:
$ cat /sys/firmware/devicetree/base/chosen/u-boot,version
2024.07-phbvwgy2ih3k0qyrjhqaxq4ahw2aa08w