CI: Build and cache artifacts on flake lock update
Initial CI building and cache pushing.  It only covers flake updates, so
any time changes are pushed to the repo the workflow will need to be
triggered with "CI_FORCE_REFRESH" and there will be a delay in cached
objects being available.  Regardless, perfect is the enemy of good and
this is good; this will serve as a point to start advancing the CI
tooling.

Tailscale is in use due to GitHub not having IPv6 connectivity...
ARM runners are only for "Organisations" only.