Add outputs.sdImages for building bootstrap images
The outputs.sdImages is structured such that:
sdImages.${buildHost}.sdImage-${manufacturer}-${model} exist.

Building an aarch64-linux/riscv image on a platform that is not the same
as the board platform will be exceptionally resource intensive as no
cache will be used.  Additionally not all packages cross-compile
successfully, such as gitFull, but it should be fine with the minimal
bootstrap images.

If you want to continue to cross compile after you've flashed a
bootstrap image, configure the SBC in your nixosConfiguration as per the
readme and add the following to your device configuration, changing the
values as required.

```
nixpkgs.hostPlatform.system = "aarch64-linux";
nixpkgs.buildPlatform.system = "x86_64-linux";
```

The sdImages are not in packages.${architecture} as evaluating them is
fairly slow, and as it evaludates multiple full-system closures it
requires a significant amount of memory.  In a nix-repl, evaluating
x86_64 and aarch64 packages consumed 22GB of memory.

Co-authored-by: Sachi King <nakato@nakato.io>