BPiR3: Add modules to minimal kernel
Add a whole slew of NFT modules, these are routers after all.  Add
ipset, as while I'm all in on nf_tables, if iptables is used ipsets are
very, very useful.

Add PPPoE as this is a router, those would be expected to be available.

Add NFT_XFRM so nftables rules load following nixpkgs commit
3c12ef3f219c1a0f458d72e7b460782287974bbd which uses "meta ipsec" in a
rule.