BPiR4: Build without autoModules
autoModules answers every kernel config option with "M" if module is an
option.  This makes the kernel extremely large and the compile times
outlandish as everything from AMDGPU to ZISOFS is built.

As the kernel currently in use has a defconfig tailored to the BPiR4 we
need to do very little to produce a functional system.  Common config is
left to turn on as much as it can find, minus a few exclusions, and we
add BTRFS support.