Bootstrap: Use new mkfs.btrfs subvol option during build
The 6.11 release of btrfs-progs will be able to make subvols during
mkfs.  Take advantage of that so as to not need to do a bunch of
filesystem manipulation during first-boot, and remove the need to
mount the filesystem on boot to inspect it on every boot.

The inline btrfs-progs override is horrible, but it will probably live
less than a month, and I do not want to override the btrfs-progs in
nixpkgs.

For now the subvols are hard-coded, as they previously were with RAMIFY.
The hard-coding will be removed in a follow up.