BPiR4: Default RTC to disabled
The RTC requires a battery to be connected to the board for it to be
useful, or in this case even usable.  As it requires additional hardware
to be useful, disable it by default.  This means that if a user attaches
and configures a different RTC, that working RTC will appear as the only
one.

By default, the RTC on the BPiR4 helpfully refuses to read if it can
tell there isn't a battery available and charged, which is better than
most.

[  144.253664] rtc-pcf8563 2-0051: low voltage detected, date/time is not reliable.

Enabling the on-board RTC is as simple as:
```
sbc.board.rtc.devices.pcf8563.enable = true;
```