Bootstrap: unblock btrfs building on systems without USER_NS
Some systems are incapable of using user namespaces in the nix build
environment, so the unshare method of making the files be owned as root
is not a method that can stay as it breaks build for those systems.

To work aruond this for now, I am adding a patch that throws away the
original uid and gid and replaces it with hard-coded uid/gid 0, or root.

This method of working around this issue is less than ideal, and the
cause of this issue is likely the same as the reason `faketime` is not
functioning.

The issues originates in btrfs-progs commit
c6464d3f99ed1dabceff1168eabb207492c37624, which switches the directory
traversal from a bespoke implementation to using a standard library
function, nftw, from libc.

As I understand it, both fakeroot and faketime use LD_PRELOAD to
carry out their modifications, and as `nftw()` is part of libc, the
`stat()` calls nftw makes, and uses to pass a stat structure to the
callback btrfs provided it, is made using a direct call rather than a
call through the linker, meaning they do not have the opportunity to
hook those function calls.