BPiR4: disable systemd TPM2 support
When using a systemd based stage-1, the systemd instance is configured
to enable TPM support by default and this by extension makes NixOS add
the TPM kernel modules to the list of modules to include in the initrd.

As this hardware does not have a TPM available, TPM modules are not
built, and thus not available.  Disable TPM support to remove this
dependnecy and allow initrd builds to succeed.

Related https://github.com/NixOS/nixpkgs/issues/344963

Signed-off-by: Sefa Eyeoglu <contact@scrumplex.net>
Signed-off-by: Sachi King <nakato@nakato.io>