Globally disable systemd TPM2 support in stage-1
This isn't a device-specific problem, very few, if any, SBCs have a TPM,
and ones that do can be explicitly enabled.  As such, it is pointless
for this to be enabled even if the system has the kernel modules, so
disable it by default with nixos-sbc usage.

This improves consistency between boards, and reduces the cycle of
systemd stage-1 being broken, then this being copied to the board.
Boards that actually have a TPM can enable it, as there will be few of
those.

Note: I do not test systemd based stage-1's.