BPiR4: Add mediatek,eeprom-data mac-address editing
As the mt7996e driver does not use of_override in a manner that works
per interface, there is no way to provide individual mac addresses via a
mac-address property in the device tree.  There is however only one
eeprom for all interfaces, and the driver loads macaddr from various
offsets inside that, so take the same approach as BPiR3 and edit macaddr
in mediatek,eeprom-data.